+++ background.js
@@ -1,2 +1,2 @@
-// Makes it so that the action icon will open sidePanel.html in the side panel.
+// The action icon will open sidePanel.html in the side panel.
 chrome.sidePanel.setPanelBehavior({ openPanelOnActionClick: true })